refactor(core): unify processing table blocks and entities
- 将粉碎台、过筛台及冲压平台方块及其方块实体统一继承ProcessingTableBlock和ProcessingTableBlockEntity
- 大幅度简化各个方块和方块实体的代码，移除水日志支持、碰撞盒和交互形状、多余的自动物品插入逻辑
- 保留方块实体的基本存储和处理功能，移除重复和复杂的物品处理实现
- 优化代码结构，提升代码的可维护性和可扩展性
